Create TEMPLATES IN FLASK
How to utilize templates with Flask using Python. 
The code shows the main functions of templates, initializations, inheritances.
There are HTML files for these codes that have to be used concurrently with these templates.
Templates are files that contain static data as well as placeholders for dynamic data. A template is rendered with specific data to produce a final document. Flask uses the Jinja template library to render templates. In your application, you will use templates to render HTML which will display in the user's browser.